+++ organization_token.go
@@ -22,6 +22,9 @@ type OrganizationTokens interface {
 	// Create a new organization token, replacing any existing token.
 	Create(ctx context.Context, organization string) (*OrganizationToken, error)
 
+	// CreateWithOptions a new organization token with options, replacing any existing token.
+	CreateWithOptions(ctx context.Context, organization string, options OrganizationTokenCreateOptions) (*OrganizationToken, error)
+
 	// Read an organization token.
 	Read(ctx context.Context, organization string) (*OrganizationToken, error)
 
@@ -41,16 +44,29 @@ type OrganizationToken struct {
 	Description string    `jsonapi:"attr,description"`
 	LastUsedAt  time.Time `jsonapi:"attr,last-used-at,iso8601"`
 	Token       string    `jsonapi:"attr,token"`
+	ExpiredAt   time.Time `jsonapi:"attr,expired-at,iso8601"`
+}
+
+// OrganizationTokenCreateOptions contains the options for creating an organization token.
+type OrganizationTokenCreateOptions struct {
+	// Optional: The token's expiration date.
+	// This feature is available in TFE release v202305-1 and later
+	ExpiredAt *time.Time `jsonapi:"attr,expired-at,iso8601,omitempty"`
 }
 
 // Create a new organization token, replacing any existing token.
 func (s *organizationTokens) Create(ctx context.Context, organization string) (*OrganizationToken, error) {
+	return s.CreateWithOptions(ctx, organization, OrganizationTokenCreateOptions{})
+}
+
+// CreateWithOptions a new organization token with options, replacing any existing token.
+func (s *organizationTokens) CreateWithOptions(ctx context.Context, organization string, options OrganizationTokenCreateOptions) (*OrganizationToken, error) {
 	if !validStringID(&organization) {
 		return nil, ErrInvalidOrg
 	}
 
 	u := fmt.Sprintf("organizations/%s/authentication-token", url.QueryEscape(organization))
-	req, err := s.client.NewRequest("POST", u, nil)
+	req, err := s.client.NewRequest("POST", u, &options)
 	if err != nil {
 		return nil, err
 	}